refactor: move animations.ts to shared folder
